Create separate files for models JS-644
