+++ libs/core/models/aired.ts
@@ -1,0 +1,9 @@
+/** The dates on which aired started and ended. */
+export type Aired = {
+
+	/** Date, when aired started. */
+	readonly start: Date | null;
+
+	/** Date, when aired ended. */
+	readonly end: Date | null;
+};
