+++ libs/core/models/anime-status.ts
@@ -1,0 +1,6 @@
+/** Anime status. */
+export enum AnimeStatus {
+	Airing = 'Airing',
+	Finished = 'Finished',
+	NotYetAired = 'Not yet aired',
+}
