+++ libs/core/models/anime-type.ts
@@ -1,0 +1,11 @@
+/** Anime type. */
+export enum AnimeType {
+	Movie = 'Movie',
+	Music = 'Music',
+	ONA = 'ONA',
+	OVA = 'OVA',
+	Promotional = 'Promotional videos',
+	Special = 'Special',
+	TV = 'TV',
+	Unknown = 'Unknown',
+}
